Separate Option to its own component
